+++ R/theme_gtsummary_roche.R
@@ -51,12 +51,12 @@ theme_gtsummary_roche <- function(font_size = NULL,
       list(
         "tbl_summary-fn:percent_fun" = label_roche_percent(),
         "pkgwide-fn:pvalue_fun" = label_roche_pvalue(),
+        "pkgwide-fun:pre_conversion" = function(x) {
+          # This runs on every gtsummary table before printing - added for tbl_strata(.header = "{strata}")
+          crane::modify_header_rm_md(x, md = "bold", type = "star")
+        },
         "add_overall.tbl_summary-arg:col_label" = "All Participants  \n(N = {style_roche_number(N)})",
         "pkgwide-str:print_engine" = print_engine,
-        "tbl_strata-fn:addnl-fn-to-run" =
-          \(x) {
-            crane::modify_header_rm_md(x, md = "bold", type = "star")
-          },
         "tbl_hierarchical-fn:addnl-fn-to-run" =
           \(x) {
             gtsummary::remove_footnote_header(x) |>